+++ client/boardspace-codename1/boardspace core/lib/ChatWidget.java
@@ -1012,7 +1012,7 @@ public class ChatWidget
 	public boolean activelyScrolling() {
 		return(activePane().activelyScrolling());
 	}
-	public boolean doMouseWheel(int xx,int yy,int amount)
+	public boolean doMouseWheel(int xx,int yy,double amount)
 	{	return activePane().doMouseWheel(xx,yy,amount);
 
 	}
