+++ client/boardspace-java/boardspace-core/online/game/GameLog.java
@@ -760,7 +760,7 @@ public class GameLog implements Opcodes
         * @param amount
         * @return
         */
-       public boolean doMouseWheel(int ex,int ey,int amount)
+       public boolean doMouseWheel(int ex,int ey,double amount)
        {	return scrollbar.doMouseWheel(ex,ey,amount);
        }
